requirements.txt
pandas==1.4.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
